+++ resumarble-core/src/main/kotlin/resumarble/core/domain/resume/facade/InterviewQuestionFacade.kt
@@ -1,21 +1,13 @@
 package resumarble.core.domain.resume.facade
 
-import jakarta.annotation.PreDestroy
-import kotlinx.coroutines.CoroutineExceptionHandler
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.cancel
-import kotlinx.coroutines.launch
 import resumarble.core.domain.gpt.ChatCompletionRequest
 import resumarble.core.domain.gpt.OpenAiMapper
 import resumarble.core.domain.gpt.application.OpenAiService
 import resumarble.core.domain.prediction.facade.PredictionFacade
-import resumarble.core.domain.prediction.facade.SavePredictionCommand
 import resumarble.core.domain.prompt.application.PromptResponse
 import resumarble.core.domain.prompt.application.PromptService
 import resumarble.core.domain.prompt.domain.PromptType
 import resumarble.core.global.annotation.Facade
-import resumarble.core.global.util.loggingErrorMarking
 import resumarble.core.global.util.loggingStopWatch
 
 @Facade
@@ -25,7 +17,6 @@ class InterviewQuestionFacade(
     private val openAiMapper: OpenAiMapper,
     private val predictionFacade: PredictionFacade
 ) {
-    private val scope = CoroutineScope(Dispatchers.IO)
 
     fun generateInterviewQuestion(command: InterviewQuestionCommand): InterviewQuestionResponse {
         val completionResult = loggingStopWatch {
@@ -34,17 +25,9 @@ class InterviewQuestionFacade(
             requestChatCompletion(completionRequest)
         }
 
-        asyncSavePrediction(openAiMapper.completionToSavePredictionCommand(command, completionResult))
+        predictionFacade.savePrediction(openAiMapper.completionToSavePredictionCommand(command, completionResult))
 
         return completionResult
-    }
-
-    private fun asyncSavePrediction(command: SavePredictionCommand) {
-        scope.launch(handler) {
-            predictionFacade.savePrediction(
-                command
-            )
-        }
     }
 
     private fun prepareCompletionRequest(
@@ -61,19 +44,7 @@ class InterviewQuestionFacade(
         )
     }
 
-    @PreDestroy
-    fun cleanUp() {
-        scope.cancel()
-    }
-
-    val handler = CoroutineExceptionHandler { _, throwable ->
-        loggingErrorMarking {
-            SAVE_PREDICTION_ERROR_MESSAGE + "${throwable.message}"
-        }
-    }
-
     companion object {
         private const val PROMPT_LANGUAGE = "korean"
-        private const val SAVE_PREDICTION_ERROR_MESSAGE = "면접 예상 질문 저장이 실패했습니다. 예외 메시지: "
     }
 }
